Added files via upload
This is now working. It takes a file called "last_modified.txt" and stores a list of modified files since last sync. It will also update itself when a file is moved or deleted. It scan the root directory (wherever this file is located) and all subfolders and files and store that list into that file. When all is done it will return a list of updated files and deleted files.